AI_chatbot/requirements.txt
Flask
Flask-cors
google-generativeai
python-dotenv
